try to fix migration error

+++ db/migrate/20161016170918_addarraytofollow_up.rb
@@ -1,6 +1,8 @@
 class AddarraytofollowUp < ActiveRecord::Migration[5.0]
 
     def change
-      remove_column :follow_ups, :checklist
+      if column_exists? :follow_ups, :checklist
+        remove_column :follow_ups, :checklist
+      end
     end
   end
